requirements.txt
bcrypt==3.1.1
cffi==1.8.3
click==6.6
cryptography==1.5.2
Flask==0.11.1
Flask-Cors==3.0.2
Flask-HTTPAuth==3.2.1
idna==2.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
mysqlclient==1.3.8
pyasn1==0.1.9
pycparser==2.17
PyJWT==1.4.2
redis==2.10.5
six==1.10.0
Werkzeug==0.11.11
